Don't skip the query field on a realtime multi query
There is no documented reason to not add the query field to the varlist
returned by a realtime multi query, despite the config category being
set to its value. Of course, there is no documentation that the category
should be set to the value either. There is lots of no documentation
when it comes to realtime. But, other engines do not skip this field so
I am forcing this backend to follow the convention, because not doing so
is very silly.
........

Merged revisions 340662 from http://svn.asterisk.org/svn/asterisk/branches/1.8
........

Merged revisions 340663 from http://svn.asterisk.org/svn/asterisk/branches/10


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@340665 f38db490-d61c-443f-a65b-d21fe96a405b