Put winning FollowMe outgoing call on hold if the caller put it on hold.
The FollowMe caller call leg is usually answered and listening to MOH.
The caller could put the call on hold while FollowMe is looking for a
winner.  The winning outgoing call is now immediately placed on hold if
the caller has put the call on hold before the winning call was selected.


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@365829 f38db490-d61c-443f-a65b-d21fe96a405b