live_ast: don't set working directory
contrib/scripts/live_ast currently assumes that it is being run from the
top-level directory of the source tree. It creates a script that will
also set the working directory.

This fix avoids the need to set the working directory if the caller sets
LIVE_AST_BASE_DIR instead.

It relies on realpath for that. If realpath is not available, it will
fall back to the original behaviour.

Review: https://reviewboard.asterisk.org/r/2027/

git-svn-id: http://svn.digium.com/svn/asterisk/trunk@370048 f38db490-d61c-443f-a65b-d21fe96a405b