Fix a crash in pjnath when starting an ICE connectivity check and immediately destroying the ICE session.
The initial ICE connectivity check is scheduled as a timer item that is to be executed immediately. It is possible for this timer item to start executing while the ICE session it is working on is destroyed. To reduce the chance of this any timer items that need to be immediately executed will be executed within the thread that has started the initial ICE connectivity check.


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@370177 f38db490-d61c-443f-a65b-d21fe96a405b