This change adds a SIP peer configuration feature to allow the peer's configured codecs to take precedence on an outgoing call.
This change introduces a new peer configuration property named
'ignore_requested_pref' that causes the requested codec to be ignored when
determining the preferred codec for an outgoing call leg. The consequence is
that Asterisk's usual efforts to prefer avoiding transcoding can be overridden
on a peer-by-peer basis where appropriate. 


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@377971 f38db490-d61c-443f-a65b-d21fe96a405b