Removed #if checks for crazy old versions of OS X.
The <arpa/nameser_compat.h> was introduced way back in OS X Panther, which
itself was end-of-lifed back in 2007. We can assume that any OS X machine
we build on will need that header file :-)

Why bother removing it? The flag we're checking (__APPLE_CC__) is actually
Apple's build number. Self-compiled versions of GCC (such as installing the
latest version of GCC from homebrew) sets the value to 0, making it useless
for this sort of compile flaggery.


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@388045 f38db490-d61c-443f-a65b-d21fe96a405b