res_stasis: remove call to missing function control_continue.
In the shuffling around of res_stasis, control_continue was renamed to
stasis_app_control_continue, but the call in res_stasis wasn't updated.
In looking into it, it turns out it wasn't really the right thing to do
in res_stasis anyways.

This patch changes the handling of received a AST_CONTROL_HANGUP frame
to be the same as receiving a NULL frame, and removed the declaration of
control_continue(), since it doesn't exist any more.

(closes issue ASTERISK-22292)
Reported by: Denis Smirnov


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@397295 f38db490-d61c-443f-a65b-d21fe96a405b