Fix UPGRADE.txt Due To Merging From Branch 11
When merging in the patch for ASTERISK-22728, the UPGRADE.txt file was changed
incorrectly.  That change should have gone into ASTERISK-11.txt.

This commit is to fix that.

Also, another comment in the UPGRADE-11.txt was missing and this commit adds
that as well.
........

Merged revisions 402115 from http://svn.asterisk.org/svn/asterisk/branches/12


git-svn-id: http://svn.digium.com/svn/asterisk/trunk@402117 f38db490-d61c-443f-a65b-d21fe96a405b